Rename files, clean up.
